Domains, Variables, Fields, and Data arrays.

A Model contains Domains, each of which contain Variables defining Fields which contain Data arrays, and Reactions with ReactionMethods that operate on the Fields to calculate model time evolution.

Model

PALEOboxes.Model — Type

Model

A biogeochemical model consisting of Domains, created from a YAML configuration file using create_model_from_config.

Domains

PALEOboxes.Domain — Type

Domain

A model region containing Variables and Reactions that act on them.

Grids

PALEOboxes.AbstractMesh — Type

AbstractMesh

Defines additional geometric and topological information for Domain

PALEOboxes.Grids.UnstructuredVectorGrid — Type

UnstructuredVectorGrid

Minimal Grid for a Vector Domain, defines only some named cells for plotting

PALEOboxes.Grids.UnstructuredColumnGrid — Type

UnstructuredColumnGrid

Minimal Grid for a Vector Domain composed of columns (not necessarily forming a 2-D array).

Fields

ncells::Int total number of cells in this Domain

Icolumns::Vector{Vector{Int}}: Icolumns[n] should be the indices of column n, in order from surface to floor, where n is also the index of

any associated boundary surface.

z_coords::Vector{FixedCoord}: z coordinates of cell mid point, lower surface, upper surface

columnnames::Vector{Symbol}: optional column names

PALEOboxes.Grids.CartesianLinearGrid — Type

CartesianLinearGrid

nD grid with netcdf CF1.0 coordinates, using Vectors for PALEO internal representation of Variables, with a mapping linear indices <–> some subset of grid indices.

The linear indices mapping should be set with set_linear_index. Conversion of Field values between the PALEO internal representation (a Vector with a linear index) and a Cartesian Array with multiple dimensions (for import and export of model output) is then implemented by cartesian_to_internal and internal_to_cartesian.

Fields

ncells::Int64: number of cells in Domain = length(linear_index) (may be subset of total points in prod(dims))

dimnames::Vector{String}: names of dimensions (ordered list)

dims::Vector{Int}: sizes of dimensions (ordered list)

coords::Vector{Vector{Float64}}: attached cell-centre coordinates for each dimension (ordered list)

coords_edges::Vector{Vector{Float64}}: attached cell-edge coordinates for each dimension (ordered list)

PALEOboxes.Grids.CartesianArrayGrid — Type

CartesianArrayGrid

nD grid with netcdf CF1.0 coordinates, using n-dimensional Arrays for PALEO internal representation of Variables

Fields

ncells::Int64: number of cells in Domain = length(linear_index) (may be subset of total points in prod(dims))

dimnames::Vector{String}: names of dimensions (ordered list)

dims::Vector{Int}: sizes of dimensions (ordered list)

coords::Vector{Vector{Float64}}: attached cell-centre coordinates for each dimension (ordered list)

coords_edges::Vector{Vector{Float64}}: attached cell-edge coordinates for each dimension (ordered list)

Subdomains

PALEOboxes.AbstractSubdomain — Type

AbstractSubdomain

Defines relationship between two Domain by mapping indices in one Domain to related indices in the other, eg interior cells adjacent to a boundary.

PALEOboxes.Grids.BoundarySubdomain — Type

BoundarySubdomain

A 2D subdomain corresponding to the 2D boundary Domain associated with a 3D interior Domain:

indices[ibnd] is the index of the 3D interior cell corresponding to a 2D boundary cell ibnd.

PALEOboxes.Grids.InteriorSubdomain — Type

InteriorSubdomain

A 3D subdomain corresponding to the 3D interior Domain associated with a 2D boundary Domain: indices[iint] is either:

missing if iint is the index of an interior cell in the 3D Domain, or

the index of the 2D boundary Domain cell corresponding to the boundary-adjacent 3D interior Domain cell iint

PALEOboxes.Grids.set_subdomain! — Function

set_subdomain!(grid::PB.AbstractMesh, subdomainname::AbstractString, subdom::PB.AbstractSubdomain, allowcreate::Bool=false)

Set Subdomain

PALEOboxes.Grids.subdomain_view — Function

subdomain_view(values, subdomain::BoundarySubdomain) -> view

Create a view on values in an interior Domain to access cells corresponding to indices in a boundary Domain.

subdomain_view(values, subdomain::InteriorSubdomain) -> var

Return unmodified values from a boundary Domain, indices to access from interior supplied by subdomain_indices

fallback when subdomain == nothing

PALEOboxes.Grids.subdomain_indices — Function

subdomain_indices(subdomain::BoundarySubdomain) -> nothing

No additional indices required to access Variables in an interior Domain from a boundary Domain (view created by subdomain_view is sufficient).

subdomain_indices(subdomain::InteriorSubdomain) -> indices

Return indices to access Variables in a boundary Domain from interior Domain (will have missing entries where interior cells do not correspond to boundary)

fallback when subdomain == nothing

Dimensions and Coordinates

Grids may define name dimensions and attach coordinates for the convenience of output visualisation. Any coordinate information required by Reactions should be supplied as Variables.

PALEOboxes.NamedDimension — Type

NamedDimension

A named dimension, with optional attached fixed coordinates coords

PALEO convention is that where possible coords contains three elements, for cell midpoints, lower edges, upper edges, in that order.

PALEOboxes.FixedCoord — Type

FixedCoord

A fixed (state independent) coordinate

Variables

PALEO Variables exist in Domains and represent biogeochemical or other quantities. They are defined by Reactions as VariableReactions which are then linked to create VariableDomains, and contain Fields which represent data (a subtype of AbstractData) in a function space (a subtype of AbstractSpace) with dimensions defined by the Domain grid (a subtype of AbstractMesh)

Dataflow dependency between Variables is represented by two pairings:

VariableReactions of VariableType VT_ReactProperty and VT_ReactDependency which are linked to create VariableDomPropDeps

VariableReactions of VariableType VT_ReactTarget and VT_ReactContributor which are linked to create VariableDomContribTargets.

PALEOboxes.VariableBase — Type

VariableBase

A Model biogeochemical Variable. Reactions access Variables using derived Types VariableReaction which are links to VariableDomains.

PALEOboxes.VariableType — Type

@enum VariableType

Enumeration of VariableBase subtypes. Allowed values:

VariableReaction: VT_ReactProperty, VT_ReactDependency, VT_ReactContributor, VT_ReactTarget

VariableDomain : VT_DomPropDep, VT_DomContribTarget

PALEOboxes.VariableDomain — Type

VariableDomain

Host (Domain) model variable.

See also: VariableDomPropDep, VariableDomContribTarget

PALEOboxes.VariableDomPropDep — Type

VariableDomPropDep

Model (Domain) VariableDomain linking a single VariableReaction{VT_ReactProperty} to multiple VariableReaction{VT_ReactDependency}.

PALEOboxes.VariableDomContribTarget — Type

VariableDomContribTarget

Model (Domain) VariableDomain linking a single VariableReaction{VT_ReactTarget} to multiple VariableReaction{VT_ReactContributor} and VariableReaction{VT_ReactDependency}.

PALEOboxes.VariableReaction — Type

VariableReaction{T}

Reaction view on a model variable.

Reactions define AbstractVarLists of VariableReactions when creating a ReactionMethod. These are used to create to VariableDomain variables when the model is initialised, linking together VariableReactions with the same name. views on Domain data Arrays are then passed to the ReactionMethod function at each timestep.

The Type parameter T is one of VT_ReactProperty, VT_ReactDependency, VT_ReactContributor, VT_ReactTarget.

localname identifies the VariableReaction within the Reaction, and can be used to set variable_attributes: and variable_links: in the .yaml configuration file.

linkreq_domain.linkreq_subdomain.linkreq_name defines the Domain, Subdomain and name for run-time linking to VariableDomain variables.

Not created directly - see CreateVariableReaction

Variable Attributes

PALEOboxes.Attribute — Type

Attribute{T}

Definition for Variable attribute name. Defines a data type T, a default_value, required (true if always present ie added with default_value when Variable is created), units, and an optional description.

Note that Variable attributes are stored as a per-Variable Dict(name => value), these definitions are only used to provide defaults, check types, and provide descriptive metadata.

ParseFromString should usually be false: a value of Type T is then required when calling set_attribute!. If ParseFromString is true, then set_attribute! will accept an AbstractString and call Base.parse(T, strvalue) to convert to T. This allows eg an enum-valued Attribute to be defined by Attribute{EnumType, true} and implementing parse(EnumType, rawvalue::AbstractString)

PALEOboxes.StandardAttributes — Constant

StandardAttributes

List of standard Variable attributes.

PALEOboxes.get_attribute — Function

get_attribute(var::VariableBase, name::Symbol, missing_value=missing) -> value

Get Variable attribute.

PALEOboxes.num_components — Function

num_components(field_data::Type{<:AbstractData}) -> Int

get number of components (optional - implement if field_data has a representation as components)

num_components(var::VariableBase) -> Int

Number of components in var.

PALEOboxes.VariableFunction — Type

@enum VariableFunction

Allowed values of :vfunction Variable Attribute, defining the Variable function to the host ODE or DAE solver.

Explicit ODE problems with dS/dt = F(S) consist of pairs of S::VFStateExplicit, F::VFDeriv Variables.

An implicit ODE problem with dU/dt = F(U) where Total variables U are functions U(S) of State variables S will consist of pairs of U::VFTotal and F::VFDeriv Variables, and also the same number of S::VF_State (in no particular order).

Algebraic constraints C(S) = 0 include variables C::VFConstraint and the same number of S::VFState, with no corresponding VF_Deriv.

Not all solvers support all combinations.

PALEOboxes.VariablePhase — Type

@enum VariablePhase

Allowed values of :vphase Variable Attribute, defining the component phase this Variable belongs to for multiphase cells.

TODO standard Variable names.

Use NetCDF CF standard names https://cfconventions.org/Data/cf-standard-names/76/build/cf-standard-name-table.html ?

Fields

Field, data and function space are defined by Variable Attributes in combination with an AbstractMesh

:field_data and optionally :data_dims define the subtype of AbstractData

:space defines the subtype of AbstractSpace

Examples:

:field_data = ScalarData, :space = ScalarSpace defines a Domain scalar (0D) quantity.

:field_data = IsotopeLinear, :space = ScalarSpace defines a Domain scalar (0D) isotope quantity.

:field_data = ScalarData, :space = CellSpace defines a per-cell quantity in a spatial Domain.

:field_data = ArrayScalarData, :data_dims =("wgrid",) :space = ScalarSpace defines a Domain scalar (0D) quantity on a wavelength grid, a Vector of length given by the value of the Domain "wgrid" data dimension (see set_data_dimension!)

PALEOboxes.Field — Type

Field{D <: AbstractData, S <: AbstractSpace, V, N, M}

A Field of data type D defined on function space S over mesh and (optionally) data_dims

PALEOboxes.get_field — Function

get_field(obj, ...) -> Field

Get Field from PALEO object obj

Spaces

PALEOboxes.AbstractSpace — Type

AbstractSpace

Defines a function space within a Domain, on a mesh defined by a Grid

PALEOboxes.ScalarSpace — Type

ScalarSpace <: AbstractSpace

A Domain position-independent quantity

PALEOboxes.CellSpace — Type

CellSpace <: AbstractSpace

A per-cell quantity. Use as Variable attribute :space to create a Variable with data array dimensions from Grid

Data types

PALEOboxes.AbstractData — Type

AbstractData

Defines a Data type that can be composed with a Space to form a Field

PALEOboxes.UndefinedData — Type

UndefinedData <: AbstractData

Undefined data type (no methods implemented). Used to indicate that a Variable can link to any data type.

PALEOboxes.ScalarData — Type

ScalarData <: AbstractData

A Field scalar (eg a biogeochemical concentration)

PALEOboxes.ArrayScalarData — Type

ArrayScalarData <: AbstractData

An Array of Field scalars (eg a intensity on a wavelength grid).

NB: only implements minimal methods sufficient to allow use as a model internal variable, not as a state variable.

Isotopes

PALEOboxes.AbstractIsotopeScalar — Type

AbstractIsotopeScalar <: AbstractData

An IsotopeScalar represents a quantity or flux with isotopic composition. Can be added, subtracted, multiplied by a scalar, and decomposed into components with the same (bulk) transport properties.

Implementation

Each IsotopeScalar should be added to IsotopeTypes, and implement:

get_total(is::IsotopeScalar) -> total

arithmetic operations +/- (ie IsotopeScalars can be added and subtracted) and *number, /number (IsotopeScalars can be scaled by a real number).

methods of the AbstractData interface

and optional isotope-specific functions, eg for a single isotope:

isotope_totaldelta(::Type{<: IsotopeScalar}, total, delta) -> IsotopeScalar()

get_delta(is::IsotopeScalar) -> delta

PALEOboxes.IsotopeLinear — Type

IsotopeLinear <: AbstractIsotopeScalar

Linearized representation of isotopic composition, where moldelta = total * delta.

PALEOboxes.isotope_totaldelta — Function

isotope_totaldelta(::Type{IsotopeLinear}, total, delta) -> IsotopeLinear

Create an IsotopeLinear from total and delta

Examples:

julia> a = PB.isotope_totaldelta(PB.IsotopeLinear, 10.0, -2.0)
(v=10.0, v_moldelta=-20.0, ‰=-2.0)

julia> b = a*2
(v=20.0, v_moldelta=-40.0, ‰=-2.0)

julia> c = a + b
(v=30.0, v_moldelta=-60.0, ‰=-2.0)

julia> PB.get_total(c)
30.0

julia> PB.get_delta(c)
-2.0

PALEOboxes.get_total — Function

generic get_total for non-isotope variable

get_total(is::IsotopeLinear) -> T

Get isotope delta (per mil)

PALEOboxes.get_delta — Function

get_delta(is::IsotopeLinear) -> T

Get isotope delta (per mil)