Reaction API

API calls used by AbstractReaction implementations.

Implementations should create a subclass of AbstractReaction containing a ReactionBase and Parameters, register a reaction factory, and implement callbacks to define VariableReactions and ReactionMethods.

Reaction struct

PALEOboxes.AbstractReaction — Type

AbstractReaction

Base Type for Reactions.

Implementation

Derived types should include a field base::ReactionBase

PALEOboxes.ReactionBase — Type

ReactionBase

Base Type for a biogeochemical Reaction.

Implementation

Include as field base in types derived from AbstractReaction

Parameters

PALEOboxes.AbstractParameter — Type

AbstractParameter

Base Type for Parameters

See also: Parameter, VecParameter

PALEOboxes.Parameter — Type

Parameter{T, ParseFromString}

A reaction parameter of type T. Create using short names ParDouble, ParInt, ParBool, ParString. Read value as <par>.v::T, set with setvalue!.

Parameters with external=true may be set from the Model-level Parameters list, if name is present in that list.

ParseFromString should usually be Nothing: a value of Type T is then required when calling setvalue!. If ParseFromString is not Nothing, then setvalue! will accept an AbstractString and call Base.parse(ParseFromString, strvalue). This allows eg an enum-valued Parameter to be defined by Parameter{EnumType, EnumType} and implementing parse(EnumType, rawvalue::AbstractString)

PALEOboxes.VecParameter — Type

VecParameter{T, ParseFromString}

A reaction parameter of type Vector{T}. Create using short names ParDoubleVec, ParStringVec. Values are <par>.v::Vector{T}. Set using standard yaml syntax for a vector eg [1, 2, 3]

ParseFromString should usually be false (see Parameter).

PALEOboxes.VecVecParameter — Type

VecVecParameter{T, ParseFromString}

A reaction parameter of type Vector{Vector{T}}. Create using short names ParDoubleVecVec. Values are <par>.v::Vector{Vector{T}}. Set using standard yaml syntax for a vector of vectors eg [[1, 2, 3], [4, 5, 6]]

ParseFromString should usually be false (see Parameter).

PALEOboxes.ParametersTuple — Function

ParametersTuple(parameters::AbstractParameter...) -> NamedTuple
ParametersTuple(parameters) -> NamedTuple

Create a NamedTuple of Parameters.

PALEOboxes.add_par — Function

add_par(reaction::AbstractReaction, par::AbstractParameter)
add_par(reaction::AbstractReaction, objectwithpars)

Add a single parameter or parameters from fields of objectwithpars to a new Reaction.

Not usually needed: Parameters in pars::ParametersTuplewill be added automatically, only needed if there are additional Parameters that are not members ofpars`.

PALEOboxes.setvalue! — Function

setvalue!(par::Parameter, value)

Set Parameter to value.

Optionally (if Parameter has Type parameter ParseFromString != Nothing) parse value from a String.

Registering with PALEOboxes framework

PALEOboxes.create_reaction — Method

create_reaction(ReactionType::Type{<:AbstractReaction}, base::ReactionBase) -> reaction::AbstractReaction

Create a ReactionType and set base field.

Default implementation may be overriden to eg set additional fields

PALEOboxes.find_reaction — Function

find_reaction(class::AbstractString) -> ReactionType

Look up "class" in list of Reactions from find_all_reactions, and return fully-qualified Reaction Type (including module prefixes).

PALEOboxes.find_all_reactions — Function

find_all_reactions() -> Dict{String, Type}

Use InteractiveUtils.subtypes(AbstractReaction) to find all currently loaded subtypes off AbstractReaction, and create a Dict with last part of the name of the Type as key (ie without the module prefix) and Type as value.

Any Types that generate non-unique keys (eg Module1.MyReactionType and Module2.MyReactionType) will generate a warning, and no entry will be added to the Dict (so if this Reaction is present in a config file, it will not be found and will error).

Optional initialisation callbacks to define Domain Grids and array sizes

One Reaction per domain may implement set_model_geometry to create and attach a grid, and set_data_dimension!.

PALEOboxes.set_model_geometry — Function

set_model_geometry(reaction, model)

Optional: define model Domain size, Subdomains etc (only implemented by Reactions that define grids).

PALEOboxes.set_data_dimension! — Function

set_data_dimension!(domain::Domain, dim::NamedDimension; allow_exists=false)

Define a Domain data dimension.

Variable data dimensions are then defined as Tuples of dim.names using the :data_dims Attribute.

Initialisation callbacks

PALEOboxes.register_methods! — Function

register_methods!(reaction)
register_methods!(reaction, model)

Add ReactionMethods, using add_method_setup!, add_method_initialize! add_method_do!. See also register_dynamic_methods!.

PALEOboxes.register_dynamic_methods! — Function

register_dynamic_methods!(reaction)
register_dynamic_methods!(reaction, model)

Optional: called after first variable link pass, to add ReactionMethods that depend on Variables generated by other Reactions (see register_methods!).

Defining VariableReactions

PALEOboxes.CreateVariableReaction — Function

CreateVariableReaction(T, localname => link_namestr, units, description; attributes=Tuple()) -> VariableReaction{T}
CreateVariableReaction(T, linklocal_namestr, units, description; attributes=Tuple()) -> VariableReaction{T}    
[deprecated] CreateVariableReaction(T, localname, units, description; link_namestr, attributes=Tuple()) -> VariableReaction{T}

Create a VariableReaction.

Not called directly: use short names VarProp, VarDep, VarContrib, VarTarget, VarPropScalar, VarDepScalar, VarContribScalar, VarTargetScalar.

Arguments

var_type::VariableType: one of VT_ReactProperty, VT_ReactDependency, VT_ReactContributor, VT_ReactTarget

localname::AbstractString: Reaction-local Variable name

link_namestr::AbstractString: <linkreq_domain>.[linkreq_subdomain.]linkreq_name. Parsed by parse_variablereaction_namestr to define the requested linking to Domain Variable.

linklocal_namestr::AbstractString: <linkreq_domain>.[linkreq_subdomain.]localname. Convenience form to define both localname and requested linking to Domain Variable, for the common case where linkreq_name == localname.

units::AbstractString: units ("" if not applicable)

description::AbstractString: text describing the variable

Keywords

attributes::Tuple(:attrb1name=>attrb1value, :attrb2name=>attrb2value, ...): variable attributes, see StandardAttributes, set_attribute!, get_attribute

PALEOboxes.parse_variablereaction_namestr — Function

parse_variablereaction_namestr(linkstr) 
    -> (linkreq_domain, linkreq_subdomain, linkreq_name, link_optional)

Parse a linkstr into component parts.

linkstr is of format: [(][<linkreq_domain>.][<linkreq_subdomain>.]<linkreq_name>[)]

Optional brackets ( ... ) set link_optional=true

linkreq_name may contain %reaction% which will later be substituted with <Reaction name>/

Examples:

julia> PALEOboxes.parse_variablereaction_namestr("foo")  # Common case eg for a property that should be public
("", "", "foo", false)

julia> PALEOboxes.parse_variablereaction_namestr("%reaction%foo")  # Reaction-private by default
("", "", "%reaction%foo", false)

julia> PALEOboxes.parse_variablereaction_namestr("ocean.foo")  # Request link to variable of same name in ocean Domain
("ocean", "", "foo", false)

julia> PALEOboxes.parse_variablereaction_namestr("(ocean.oceansurface.goo)") # Full syntax
("ocean", "oceansurface", "goo", true)

PALEOboxes.set_attribute! — Function

set_attribute!(var::VariableBase, name::Symbol, value, allow_create=false)

Set Variable attribute.

Defining collections of VariableReactions

PALEOboxes.AbstractVarList — Type

AbstractVarList

Variables required by a ReactionMethod methodfn are specified by a Tuple of VarList_xxx <: AbstractVarList, each containing a collection of VariableReaction.

These are then converted (by the create_accessors method) to a corresponding Tuple of collections of views on Domain data arrays , which are then be passed to the ReactionMethod methodfn.

Implementation

Subtypes of AbstractVarList should implement:

a constructor that takes a collection of VariableReactions

create_accessors, returning the views on Domain data arrays in a subtype-specific collection.

get_variables, returning the collection of VariableReactions (as a flat list).

PALEOboxes.VarList_single — Type

VarList_single(var; components=false) -> VarList_single

Create a VarList_single describing a single VariableReaction, create_accessors will then return a single accessor.

PALEOboxes.VarList_namedtuple — Type

VarList_namedtuple(varcollection; components=false) -> VarList_namedtuple

Create a VarList_namedtuple describing a collection of VariableReactions, create_accessors will then return a NamedTuple with field names = VariableReaction.localname and field values = corresponding data arrays.

If components = true, each NamedTuple field will be a Vector of data array components.

PALEOboxes.VarList_tuple — Type

VarList_tuple(varcollection; components=false) -> VarList_tuple

Create a VarList_tuple describing a collection of VariableReactions, create_accessors will then return a Tuple of data arrays.

If components = true, each Tuple field will be a Vector of data array components.

PALEOboxes.VarList_vector — Type

VarList_vector(varcollection; components=false, forceview=false) -> VarList_vector

Create a VarList_vector describing a collection of VariableReactions, create_accessors will then return a Vector of data arrays.

If components = true, each Vector element will be a Vector of data array components.

If forceview = true, each accessor will be a 1-D view to help type stability, even if this is redundant (ie no view required, v::Vector -> view(v, 1:length(v)))

PALEOboxes.VarList_vvector — Type

VarList_vvector(Vector{Vector{VariableReaction}}::vars; components=false) -> VarList_vvector

Create a VarList_vvector describing a Vector of Vectors of VariableReactions, create_accessors will then return a Vector of Vectors of data arrays.

If components = true, each Vector of Vectors element will be a Vector of data array components.

PALEOboxes.VarList_nothing — Type

VarList_nothing() -> VarList_nothing

Create a placeholder for a missing/unavailable VariableReaction. create_accessors will then return nothing.

PALEOboxes.VarList_tuple_nothing — Type

VarList_tuple_nothing(nvar) -> VarList_tuple_nothing

Create a placeholder for nvar missing/unavailable VariableReactions. create_accessors will then return an NTuple{nvar, Nothing}.

Registering ReactionMethods

PALEOboxes.ReactionMethod — Type

ReactionMethod

Defines a callback function methodfn with Variables vars, to be called from the Model framework either during setup or as part of the main loop.

Fields

methodfn::Function: callback from Model framework

reaction::AbstractReaction: the Reaction that created this ReactionMethod.

name::String: a descriptive name, eg generated from the name of methodfn

vars::Tuple{Vararg{AbstractVarList}}: AbstractVarLists of VariableReactions. Corresponding Variable accessors (views on Arrays) will be provided to the methodfn callback.

p::P: optional context field (of arbitrary type) to store data needed by methodfn.

operatorID::Vector{Int64}: operator ID (to allow operator splitting)

domain::Domain: the Domain that methodfn operates on (ie will be called with cellrange for this domain).

preparefn::Function: optional function to eg define buffers before main loop starts.

methodfn

The methodfn callback is:

methodfn(m::ReactionMethod, vardata::Tuple, cellrange::AbstractCellRange, modelctxt)

With arguments:

m::ReactionMethod: context is available as m.reaction::AbstractReaction (the Reaction that defined the ReactionMethod), and m.p (an arbitrary extra context field supplied when ReactionMethod created).

vardata: A Tuple of collections of views on Domain data arrays, defined by vars

cellrange::AbstractCellRange: range of cells to calculate.

modelctxt:

for a setup method, :initial_value or :norm_value defining the type of setup requested

for a main loop method deltat providing timestep information eg for rate throttling.

preparefn

An optional preparefn callback can be supplied eg to allocate buffers that require knowledge of the data types of vardata or to cache expensive calculations:

preparefn(m::ReactionMethod, vardata::Tuple) -> vardata

This is called after model arrays are allocated, and prior to setup.

PALEOboxes.add_method_setup! — Function

add_method_setup!(reaction::AbstractReaction, method::AbstractReactionMethod)
add_method_setup!(reaction::AbstractReaction, methodfn::Function, vars::Tuple{Vararg{AbstractVarList}}; kwargs...) -> ReactionMethod

Add or create-and-add a setup method (called before main loop) eg to set persistent data or initialize state variables. methodfn, vars, kwargs are passed to ReactionMethod.

PALEOboxes.add_method_initialize! — Function

add_method_initialize!(reaction::AbstractReaction, method::AbstractReactionMethod)
add_method_initialize!(reaction::AbstractReaction, methodfn::Function, vars::Tuple{Vararg{AbstractVarList}}; kwargs...) -> ReactionMethod

Add or create-and-add an initialize method (called at start of each main loop iteration) eg to zero out accumulator Variables. methodfn, vars, kwargs are passed to ReactionMethod.

PALEOboxes.add_method_do! — Function

add_method_do!(reaction::AbstractReaction, method::AbstractReactionMethod)
add_method_do!(reaction::AbstractReaction, methodfn::Function, vars::Tuple{Vararg{AbstractVarList}}; kwargs...) -> ReactionMethod

Add or create and add a main loop method. methodfn, vars, kwargs are passed to ReactionMethod.

Predefined ReactionMethods

Setup and initialization of Variables

PALEOboxes.add_method_setup_initialvalue_vars_default! — Function

add_method_setup_initialvalue_vars_default!(react::AbstractReaction, variables [; kwargs...])

Create and add a default method to initialize Variables matching filterfn (defaults to state Variables) to Attribute :initial_value at beginning of integration.

Keywords

filterfn: set to f(var)::Bool to override the default selection on :vfunction for state variables.

transfer_attribute_vars=[]: Set to a list of the same length as variables to initialise variables from attributes of transfer_attribute_vars.

setup_callback=(method, attribute_value, var, vardata) -> nothing: Set to a function that is called after each Variable initialisation eg to store norm values.

convertvars=[]

convertfn = (convertvars_tuple, i) -> 1.0

convertinfo = ""

Including volume etc conversion

Set convertvars to a Vector of Variables (eg for cell volume) and supply convertfn and convertinfo to initialize to :initial_value*convertfn(convertvars_tuple, i) where the argument of convertfn is the Tuple generated by VarList_tuple(convertvars).

Example: To interpret :initial_value as a concentration-like quantity:

convertvars = [volume], 
convertfn = ((volume, ), i) -> volume[i], 
convertinfo = " * volume"

PALEOboxes.add_method_initialize_zero_vars_default! — Function

add_method_initialize_zero_vars_default!(react::AbstractReaction, variables=PB.get_variables(react))

Create and add a default method to initialize Variables to zero at beginning of each timestep. Defaults to adding all Variables from react with :initialize_to_zero attribute true.

NB: TODO variables are converted to VarDep (no dependency checking or sorting needed, could define a VarInit or similar?)

Adding totals for Variables

PALEOboxes.add_method_do_totals_default! — Function

add_method_do_totals_default!(react::AbstractReaction, variables=PB.get_variables(react))

Create and add a default method to initialize Variables to zero at beginning of each timestep. Defaults to adding all Variables from react with :initialize_to_zero attribute true.

NB: TODO variables are converted to VarDep (no dependency checking or sorting needed, could define a VarInit or similar?)

Chemical reactions

PALEOboxes.RateStoich — Type

RateStoich(ratevartemplate, stoich_statevarname; deltavarname_eta=nothing,
           sms_prefix="", sms_suffix="_sms") -> RateStoich

Calculate fluxes for a biogeochemical reaction given rate, stoichiometry, and optionally isotope eta.

Add to a Reaction using create_ratestoich_method and add_method_do!.

A Property Variable should be set to provide the reaction rate (often this is implemented by another method of the same Reaction).

This method will then link to that (using the local and link names supplied by ratevartemplate) and calculate the appropriate product rates, omitting products that are not present (VariableReaction not linked) in the Model configuration.

Arguments:

ratevartemplate::Union{VarPropT, VarDepT}: used to define the rate variable local and link names.

stoich_statevarname: collection of Tuple(stoichiometry, name) eg ((-2.0, "O2"), (-1.0,"H2S::Isotope"), (1.0, "SO4::Isotope"))

deltavarname_eta: optional tuple of variable delta + eta ("SO4_delta", -30.0) or ("SO4_delta", rj.pars.delta). If a Parameter is supplied, this is read in do_react_ratestoich to allow modification.

Examples:

Create a RateStoich representing the reaction 2 O2 + H2S -> H2SO4

julia> myratevar = PALEOboxes.VarProp("myrate", "mol yr-1", "a rate");

julia> rs = PALEOboxes.RateStoich(myratevar, ((-2.0, "O2"), (-1.0,"H2S"), (1.0, "SO4")));

julia> rs.stoich_statevarname
((-2.0, "O2"), (-1.0, "H2S"), (1.0, "SO4"))

PALEOboxes.create_ratestoich_method — Function

create_ratestoich_method(reaction::AbstractReaction, ratestoich::RateStoich; isotope_data=ScalarData)
    -> ReactionMethod

Create method (see RateStoich).

Implementing Reaction method functions

Method functions should iterate over cells using the supplied cellrange.indices.

Optimising loops over cells using explicit SIMD instructions

Reactions with simple loops over cellindices that implement time-consuming per-cell calculations may be optimised by using explicit SIMD instructions.

PALEOboxes.SIMDutils.SIMDIter — Type

SIMDIter(baseiter, Val{N})
SIMDIter(baseiter, ::Type{SIMD.Vec{N, U}})
SIMDIter(baseiter, ::Val{1}) # scalar fallback
SIMDIter(baseiter, ::Type{U}) where {U <: Real} # scalar fallback

Iterator that takes up to N SIMD elements at a time from baseiter (which should represent indices into a Vector). See Julia package SIMD.jl

If baseiter contained 1 or more but less then N elements, then indices is filled with repeats of the last available element.

Returns Tuple indices (length N).

Examples

v_a = [1.0, 2.0, 3.0, 4.0, 5.0]
v_b = similar(v_a)

iter = eachindex(v_a) # iter should represent indices into a Vector

# simplest version - Float64 x 4, ie type of v_a x 4

for indvec in SIMDIter(iter, Val(4))
    x = v_a[indvec]  # x is a packed SIMD vector
    v_b[indvec] = x
end

# with type conversion - Float32 x 8, ie explicitly change Type of SIMD vector

ST = SIMD.Vec{8, Float32}}    
for indvec in SIMDIter(iter, ST)
    #   v = vec[indvec]  <--> vgatherind(ST, vec, indvec)
    #   vec[indvec] = v  <--> vscatterind!(v, vec, indvec)
    #   vec[indvec] += v <--> vaddind!(v, vec, indvec) 

    x = vgatherind(ST, v_a, indices)  # x is a packed SIMD vector with type conversion to Float32
    vscatterind!(x, v_b, indices)
end

Internal details of Variable arrays accessor generation

VariableReactions in the AbstractVarLists for a ReactionMethod are processed by create_accessor to supply views on arrays as corresponding arguments to the ReactionMethod function.

PALEOboxes.create_accessor — Function

 create_accessor(var::VariableReaction, modeldata, components, [,forceview=false])
    -> accessor or (accessor, subdomain_indices)

Creates a view on a (single) VariableDomain data array linked by var::VariableReaction. Called by an AbstractVarList create_accessors implementation to generate a collection of views for multiple VariableReactions.

Returns:

if var is linked, an accessor or Tuple (accessor, subdomain_indices) that provides a view on variable data.

if var is not linked, nothing if var is optional, or errors and doesn't return if var is non-optional.

Mapping of indices for Subdomains <–> Domains:

if no Subdomain, returns unmodified indices (if forceview=false), or an equivalent view (if forceview=true, this is to help type stability)

if Variable is a Domain interior and Subdomain is a Domain boundary, accessor is a view with a subset of Domain indices.

if Variable is a Domain boundary and Subdomain is the Domain interior, returns a Tuple with subdomain_indices, length(Subdomain size), with missing for interior points.

Mapping of multi-component (Isotope) Variables:

If components=false:

map multi-component Variable to accessor::IsotopeArray

return a single-component Variable as a accessor::AbstractArray.

If components=true:

variable data as a accessor::Vector{Array}, length=number of components

PALEOboxes.create_accessors — Function

create_accessors(va::AbstractVarList, modeldata::AbstractModelData) -> collection of views

Call create_accessor to return a collection of views on Domain data arrays for VariableReaction in va. Collection and view are determined by va Type.