Generic Reaction catalog

PALEOboxes includes a catalog of generic Reactions as a starting point for model construction.

Reservoirs

PALEOboxes.Reservoirs.ReactionReservoirScalar — Type

ReactionReservoirScalar

A single scalar biogeochemical reservoir with optional paired isotope reservoir, for use in a 0D Domain (eg sedimentary or ocean reservoirs for COPSE N. M. Bergman, Timothy M Lenton, Andrew J Watson (2004)).

Creates State and associated Variables, depending on parameter settings:

const=false, stateexplicit=true: usual case, create state variable R (units mol, with attribute vfunction=VF_StateExplicit) and R_sms (units mol yr-1, with attribute vfunction=VF_Deriv).

const=true, stateexplicit=true: a constant value, create R (a Property)

const=false, stateexplicit=false: for use with an implicit state variable, create R_sms (with attribute vfunction=VF_Deriv) and a Dependency R (which should be linked to a Variable with attribute vfunction=VF_Total calculated elsewhere, which should also set up initial value).

In addition:

a Property R_norm (normalized value) is always created.

if parameter field_data <: AbstractIsotopeScalar (eg IsotopeLinear), a Property R_delta is created.

The local name prefix R should then be renamed using variable_links: in the configuration file.

Initialisation

state_explicit=true: (the usual case) initial and norm value is set in the variable_attributes: section in the configuration file, using R:initial_value, R:initial_delta, and R:norm_value.

state_explicit=false: initial and norm values should be set by the linked implicit state Variable

Example configuration in .yaml file

            reservoir_P:  # 0D ocean Phosphorus
                class: ReactionReservoirScalar
                parameters:
                    # field_data: ScalarData        # change to IsotopeLinear to represent an isotope
                    # const: false                  # true to fix to constant value
                variable_links:
                    R*: P                           # rename to represent Phosphorus
                variable_attributes:
                    R:norm_value:           3.1e15  # mol 
                    R:initial_value:        3.1e15  # mol

See also

ReactionReservoir (one value per cell for a spatially resolved Domain eg ocean), ReactionReservoirWellMixed (one value for a whole spatially resolved Domain).

PALEOboxes.Reservoirs.ReactionReservoir — Type

ReactionReservoir,  ReactionReservoirTotal

A single (vector) reservoir (state variable) representing a biogeochemical tracer, ie one value per cell in a spatially-resolved Domain (eg ocean).

State Variables created depend on parameters:

stateexplicit=true: usual case to represent an ODE variable and time derivative, create state variable R (mol m-3, with attribute vfunction=VF_StateExplicit) and R_sms (mol m-3 yr-1, with attribute vfunction=VF_Deriv).

stateexplicit=false: for use with an ODE implicit state variable, create R_sms (with attribute vfunction=VF_Deriv) and a Dependency R (which should be linked to a Variable with attribute vfunction=VF_Total calculated elsewhere, this Variable should also set the initial value).

In addition:

a Property R_conc (concentration in mol m-3) is always created.

if parameter field_data <: AbstractIsotopeScalar (eg IsotopeLinear), a Property R_delta is created.

ReactionReservoirTotal also calculates the Domain total R_total (units mol), eg to check budgets.

Local name prefix R should then be renamed using variable_links: in the configuration file.

Initialisation

Initial value is set using variable_attributes: in the configuration file, using R:initial_value and R:initial_delta (NB: although initial_value is set on R, it sets concentration in mol m-3.)

Transport is defined by attributes :advect, :vertical_movement (m d-1) set on the concentration variable R_conc. Optical extinction is defined by the :specific_light_extinction (m^2 mol-1) attribute set on the concentration variable R_conc.

Example configuration in .yaml file

            reservoir_P:  # ocean Phosphorus
                class: ReactionReservoirTotal       # include _total (mol)
                parameters:
                    # field_data: ScalarData        # change to IsotopeLinear to represent an isotope                       
                variable_links:
                    R*: P                           # rename to represent Phosphorus
                variable_attributes:
                    R:norm_value:           3e-3    # mol m-3, normalisation value (used by some solvers)
                    R:initial_value:        2e-3    # mol m-3, initial concentration

See also

ReactionReservoirWellMixed (one value for the whole Domain), ReactionReservoirScalar (one value for a reservoir in a 0D Domain eg for COPSE N. M. Bergman, Timothy M Lenton, Andrew J Watson (2004)), ReactionReservoirConst (constant time-independent value ie no state variable).

PALEOboxes.Reservoirs.ReactionReservoirWellMixed — Type

ReactionReservoirWellMixed

A Scalar Reservoir representing a well-mixed tracer in a vector Domain (eg ocean). Provides a scalar state Variable R and R_sms, and also vector Variables: R_conc (set to uniform concentration), and R_vec_sms Target for accumulating fluxes that is added to the scalar R_sms.

Initialisation

Initial value is set in the variable_attributes: section in the configuration file, using R:initial_value and R:initial_delta. NB: may be initialized either from mean concentration or total (set by parameter initialization_type)

TODO salinity normalisation.

See also

ReactionReservoir (one value per cell), ReactionReservoirScalar (one value for a reservoir in a 0D Domain eg for COPSE N. M. Bergman, Timothy M Lenton, Andrew J Watson (2004)).

PALEOboxes.Reservoirs.ReactionReservoirConst — Type

ReactionReservoirConst

A single (vector) constant tracer R_conc (constant replacement for a ReactionReservoir).

Local name prefix R should then be renamed using variable_links: in the configuration file.

Initialisation

Set :initial_value, :initial_delta on R_conc (mol m-3) in the variable_attributes: section of the config file.

TODO salinity normalisation.

Example configuration in .yaml file

            reservoir_B:  # Constant value for ocean Boron 
                class: ReactionReservoirConst
                parameters:
                    field_data: IsotopeLinear
                variable_links:
                    R*: B
                variable_attributes:                      
                    R_conc:initial_value:       0.4269239 # contemporary value
                    R_conc:initial_delta:       34.0

See also

ReactionReservoir

PALEOboxes.Reservoirs.ReactionReservoirForced — Type

ReactionReservoirForced

A single (vector) constant tracer (constant replacement for a ReactionReservoir), with forcing.

Calculates R_conc = R_conc_initial * R_FORCE.

Local name prefix R should then be renamed using variable_links: in the configuration file.

Initialisation

NB: set :initial_value, :initial_delta on R_conc_initial in the variable_attributes: section of the config file.

TODO salinity normalisation.

See also

ReactionReservoirConst, ReactionReservoir

PALEOboxes.Reservoirs.ReactionConst — Type

ReactionConst, ReactionScalarConst

Create constant Property Variables with names from parameter constnames.

Initialisation

Constant values set by :initial_value, :initial_delta attributes in the variable_attributes: section of the configuration file.

Example configuration in .yaml file

    atmfloor:
        reactions:                
            floorstubgasmr:  # Provide mixing-ratio boundary condition for a subset of atmospheric variables
                class: ReactionConst
                parameters:
                    constnames: ["O2_mr", "CH4_mr", "CO2_mr", "H2_mr"] #, "H2O_mr"]  
                variable_attributes:        
                    O2_mr:initial_value:  [0.21]  # mol/mol
                    CH4_mr:initial_value:  [0.7443e-6] # [.NaN] # [1.271e-6] # [1.8e-6]  # mol/mol
                    CO2_mr:initial_value:  [300e-6]  # mol/mol
                    H2_mr:initial_value:   [.NaN]  # mol/mol

See also

ReactionReservoirConst, ReactionReservoirScalar. These provide additional variables (eg R_delta) to allow them to function as a drop-in replacement for a non-constant Reservoir.

Linking to Reservoirs from a Reaction

PALEOboxes.Reservoirs.ReservoirLinksVector — Function

ReservoirLinksVector(reservoirlist) -> (res::Vector, sms::Vector, diag::Vector)

Convenience function to create variables required for a Reaction to link to a list of Reservoir variables. res contains VariableReactions <reservoir_name>, sms <reservoir_name>_sms that link to State, State_sms variables. diag contains VariableReactions <reservoir_name>_norm etc that link to additional properties.

Arguments

reservoirlist::[(reservoir_name[::Isotope], units, description), ...]: list of Reservoirs

Variable Statistics

PALEOboxes.VariableStats.ReactionSum — Type

ReactionSum, ReactionVectorSum

A sum of variables (eg budget).

If Parameter component_to_add == 0, all components of Isotopes are included.

If Parameter component_to_add == component_number, a single component only is included.

PALEOboxes.VariableStats.ReactionWeightedMean — Type

ReactionWeightedMean

Weighted mean (eg by area or volume) of Variable

PALEOboxes.VariableStats.ReactionAreaVolumeValInRange — Type

ReactionAreaVolumeValInRange

Fraction of Domain area or volume with Variable in a range of values.

Fluxes

PALEOboxes.Fluxes — Module

Fluxes

Create and manage collections of Target and Contrib Variables used to define fluxes within or between Domains.

Conventions for defining biogeochemical fluxes within Domains

Particulate organic matter (including CaCO3) with stoichiometry Corg:N:P:Ccarb is usually transferred as: flux_list ["P", "N", "Corg::CIsotope", "Ccarb::CIsotope"] with a prefix indicating the function (eg export_).

Solute fluxes are usually transferred as: flux_list ["DIC::CIsotope", "TAlk", "Ca", "P", "O2", "SO4::SIsotope", "H2S::SIsotope", "CH4::CIsotope"] (where these names match the Reservoir names for the solutes), with a prefix indicating the function (usually just flux_ or soluteflux_).

Conventions for defining global fluxes between modules

The model configuration .yaml file should create a Domain for each global flux, containing one or more Fluxes.ReactionFluxTarget. Fluxes are then transferred (copied) by adding a Fluxes.ReactionFluxTransfer to each destination Domain.

Naming conventions for Earth system fluxes:

Domain name	target prefix	flux_list (illustrative, add as needed)

fluxAtoLand	flux_	["CO2::CIsotope", "O2"]

fluxRtoOcean	flux_	["DIC::CIsotope", "TAlk", "Ca", "P", "SO4::SIsotope"]

fluxOceanBurial	flux_	["Corg::CIsotope", "Ccarb::CIsotope", "Porg", "Pauth", "PFe", "P", "GYP::SIsotope", "PYR::SIsotope"]

fluxSedCrusttoAOcean	flux_	["C::CIsotope", "S::SIsotope", "Redox"]

fluxLandtoSedCrust	flux_	["Ccarb::CIsotope", "Corg::CIsotope", "GYP::SIsotope", "PYR::SIsotope"]

fluxOceanfloor	particulateflux_	["P", "N", "Corg::CIsotope", "Ccarb::CIsotope"]

soluteflux_	["DIC::CIsotope", "TAlk", "Ca", "P", "O2", "SO4::SIsotope", "H2S::SIsotope", "CH4::CIsotope"]

fluxAtmtoOceansurface	flux_	["CO2::CIsotope", "CH4::CIsotope", "O2"]

PALEOboxes.Fluxes.ReactionFluxTarget — Type

ReactionFluxTarget

Provides either a target for fluxes from fluxlist in an input Domain or a constant stub, optionally calculates totals.

For each fluxname in fluxlist, creates:

if const_stub==false, an input VarTarget pars.target_prefix.v*"flux_"*fluxname (or if const_stub==true, a constant VarProp).

if flux_totals is true, a total VarPropScalar target_prefix.v*"flux_total_"*fluxname.

PALEOboxes.Fluxes.ReactionFluxTransfer — Type

ReactionFluxTransfer

Copy fluxes, optionally using transfer matrices to define cell-cell mappings if input and output fluxes are in a different Domain.

A transfer is defined by Parameters:

input_fluxes: A list of input Variables generated by matches to string [inputdomain.][inputsubdomain.]prefix$fluxname$suffix where $fluxname$ matches any string. These have local names input_$fluxname$.

output_fluxes: Corresponding list of output Variables. [outputdomain.][outputsubdomain.]prefix$fluxname$suffix where $fluxname$ is generated from matches of input_fluxes. These have local names output_$fluxname$ and are optional (ie ignored if not linked). Usually the output Domain is the Domain hosting the ReactionFluxTransfer (in general, it must be a Domain of the same size as the Domain hosting the ReactionFluxTransfer).

transfer_multiplier: A scalar multiplier (usually 1.0).

transfer_matrix: Form of transfer matrix.

'Identity' copies fluxes directly assuming the output Domain has the same size as the input Domain.

'Distribute' uniformly distributes fluxes from input -> output, including common cases n -> 1 (input flux is summed to a scalar output), 1 -> m (scalar input is distributed uniformly to m output cells).

There are three common cases:

Transfer within a Domain, using transfer_matrix Identity

Transfer from a fluxXXX Domain to the Domain hosting the ReactionFluxTransfer, using transfer_matrix to define a mapping if these Domains are of different sizes.

Transfer from a fluxXXX Domain to the boundary cells of an interior Domain (eg ocean.oceansurface), where the Domain hosting the ReactionFluxTransfer is the corresponding boundary Domain (eg oceansurface).

Contributing to flux couplers from a Reaction

PALEOboxes.Fluxes.FluxContrib — Function

FluxContrib(
    fluxprefix::AbstractString, flux_list;
    isotope_data::Dict,
    [, space=PB.CellSpace][, alloptional=true]
) -> vartuple::NamedTuple

Create a NamedTuple of VarContrib from flux_prefix.*flux_list, needed for a Reaction to write to a flux coupler. Entries in flux_list of form fluxname::isotope are parsed to look up isotope type from isotope_data Dict.

For each fluxname in flux_list, the generated VarContrib vartuple.fluxname have link_namestr=$(fluxprefix)$(fluxname)" and localname="$(fluxprefix)$(fluxname) (with any . in fluxprefix substituted to _).

Example:

julia> fluxRtoOcean = PB.Fluxes.FluxContrib("fluxRtoOcean.flux_", ["P", "SO4::SIsotope"], isotope_data=Dict("SIsotope"=>PB.IsotopeLinear), space=PB.ScalarSpace);

julia> fluxRtoOcean.P
PALEOboxes.VariableReaction{PALEOboxes.VT_ReactContributor}
  localname='fluxRtoOcean_flux_P'
  link_name='fluxRtoOcean.flux_P[unlinked]'
  attributes=Dict{Symbol, Any}(:description => "flux P", :space => PALEOboxes.ScalarSpace, :standard_variable => false, :norm_value => 1.0, :data_dims => String[], :long_name => "", :vfunction => PALEOboxes.VF_Undefined, :field_data => PALEOboxes.ScalarData, :initial_delta => 0.0, :units => "mol yr-1"…)

julia> fluxRtoOcean.SO4
PALEOboxes.VariableReaction{PALEOboxes.VT_ReactContributor}
  localname='fluxRtoOcean_flux_SO4'
  link_name='fluxRtoOcean.flux_SO4[unlinked]'
  attributes=Dict{Symbol, Any}(:description => "flux SO4", :space => PALEOboxes.ScalarSpace, :standard_variable => false, :norm_value => 1.0, :data_dims => String[], :long_name => "", :vfunction => PALEOboxes.VF_Undefined, :field_data => PALEOboxes.IsotopeLinear, :initial_delta => 0.0, :units => "mol yr-1"…)

Forcings

PALEOboxes.Forcings.ReactionForceInterp — Type

ReactionForceInterp

F forcing interpolated from parameters (use configuration file to rename output variable F) NB: no extrapolation ! (so eg set guard values at -1e30, 1e30)

PALEOboxes.GridForcings.ReactionForceGrid — Type

ReactionForceGrid

Apply time-dependent, periodic forcing from a variable in a netcdf or Matlab file, optionally applying a scaling, a constant linear offset, and a linear offset generated from a scalar model variable.

Reads records tidx_start:tidx_end (assumed to be the last dimension) for data_var from a gridded dataset in netcdf_file or matlab_file, maps grid to linear using the Domain grid (which must match that of data_var).

If tidx_end > tidx_start (ie multiple records), reads a netcdf or Matlab variable named time_var, applies periodicity cycle_time, and uses this to linearly interpolate to model time.

Then applies forcing:

F = scale*`data_var` + constant_offset + scale_offset_var*`scalar_offset_var`

Optionally (if interp_vars is non-empty), interpolate forcing from additional dimensions in the netcdf file, given values supplied by additional Variable dependencies. NB: netcdf dimensions order must be gridvars x interpvars x timevar, where order within interpvars also must match netcdf order.

PALEOboxes.FluxPerturb.ReactionFluxPerturb — Type

ReactionFluxPerturb

Add scalar flux F interpolated from parameters (use configuration file to rename output variable F) NB: no extrapolation ! (so eg set guard values at -1e30, 1e30)

PALEOboxes.FluxPerturb.ReactionRestore — Type

ReactionRestore

Adds RestoringFlux in response to discrepancy between Variable WatchLevel and Parameter RequiredLevel

Grids

Minimal generic model grids for test purposes.

These just define the Domain size, and don't provide coordinate or metric information (cell volume, area). Models will usually require a Reaction with a Domain-specific implementation (eg for ocean, atmosphere) that defines coordinates and volumes etc and also implements transport (advection, eddy diffusion, etc).

PALEOboxes.GridReactions.ReactionUnstructuredVectorGrid — Type

ReactionUnstructuredVectorGrid

Create an PB.UnstructuredVectorGrid with ncells (from ncells Parameter).

PALEOboxes.GridReactions.ReactionCartesianGrid — Type

ReactionCartesianGrid

Create a PB.CartesianArrayGrid with dims and dimnames`

PALEOboxes.GridReactions.ReactionGrid2DNetCDF — Type

ReactionGrid2DNetCDF

Create a 2D PB.CartesianLinearGrid from grid information in a NetCDF file.