Solver API

A Model contains Domains, each of which contain Variables defining Fields, and Reactions with ReactionMethods that operate on the Fields to calculate model time evolution.

Create and initialise

PALEOboxes.create_model_from_config — Method

create_model_from_config(
    config_file::AbstractString, configmodel::AbstractString;
    modelpars::Dict = Dict()
) -> model::Model

create_model_from_config(
    config_files,
    configmodel::AbstractString;
    modelpars::Dict = Dict()
) -> model::Model

Construct model from a single YAML config_file, or from a collection of config_files, which are read in order and concatenated before being parsed as yaml.

Optional argument modelpars provides parameters that override those in <configmodel>: parameters: section.

PALEOboxes.create_modeldata — Function

create_modeldata(model::Model [, eltype] [; threadsafe]) -> modeldata::ModelData

Create a new ModelData struct for model variables of element type eltype.

PALEOboxes.ModelData — Type

ModelData{T}

Contains model data arrays of element type T.

Fields

cellranges_all::Vector{AbstractCellRange}: default cellranges covering all domains

dispatchlists_all: default dispatchlists covering all domains

solver_view_all: optional untyped context field for use by external solvers.

PALEOboxes.allocate_variables! — Function

allocate_variables!(model, modeldata; [hostdep] [,eltypemap])

Allocate memory for Domain variables for every Domain in model.

See allocate_variables!(domain::Domain, modeldata::AbstractModelData).

allocate_variables!(domain, modeldata; [hostdep=false] [, eltypemap::Dict{String, DataType}])

Allocate memory for Domain Variables.

If hostdep=false, only internal Variables are allocated, allowing host-dependent Variables (usually state Variables and derivatives + any external dependencies) to be set to views on host-managed arrays.

Element type of allocated Arrays is determined by eltype(modeldata) (the usual case, allowing use of AD types), or Variable :datatype attribute if present (allowing Variables to ignore AD types). :datatype may be either a Julia DataType (eg Float64), or a string to be looked up in eltypemap.

PALEOboxes.check_ready — Method

check_ready(model, modeldata; [throw_on_error=true] [, check_hostdep_varnames=true] [, expect_hostdep_varnames=["global.tforce"]]) -> ready

Check all variable pointers set, and no unexpected host-dependent non-state Variables (ie unlinked Variables)

PALEOboxes.check_configuration — Method

Check configuration

PALEOboxes.initialize_reactiondata! — Method

initialize_reactiondata!(model::Model, modeldata::AbstractModelData)

Processes VarList_...s from ReactionMethods and populates modeldata.sorted_methodsdata_... with sorted lists of ReactionMethods and corresponding Variable accessors.

Calls create_dispatch_methodlists(model, modeldata, modeldata.cellranges_all) to set modeldata.dispatchlists_all to (optional) default ReactionMethodDispatchLists for entire model.

PALEOboxes.dispatch_setup — Function

dispatch_setup(model, attribute_name, modeldata, cellranges=modeldata.cellranges_all)

Call setup methods, eg to initialize data arrays (including state variables).

attribute_name defines the setup operation performed. dispatch_setup should be called in sequence with attribute_name = :

:setup: initialise Reactions and set up any non-state Variables (eg model grid Variables)

:norm_value: set state Variable values from :norm_value attribute in .yaml file, and initialise any Reaction state that requires this value.

:initial_value (optional): set state Variable values from :initial_value attribute in .yaml file.

PALEOboxes.create_dispatch_methodlists — Function

create_dispatch_methodlists(model::Model, modeldata::AbstractModelData, cellranges) 
    -> (;list_initialize::ReactionMethodDispatchList, list_do::ReactionMethodDispatchList)

Compile lists of initialize and do methods + corresponding cellrange for main loop do_deriv.

Subset of methods and cellrange to operate on are generated from supplied cellranges.

PALEOboxes.ReactionMethodDispatchList — Type

ReactionMethodDispatchList

Defines a list of ReactionMethod with corresponding CellRange and views on Variable data (sub)arrays.

Attaching numerical solvers

High-level access to aggregated collections of state Variables and derivatives is provided by VariableAggregator (see Accessing model objects for low-level access).

PALEOboxes.VariableAggregator — Type

VariableAggregator{T, F, C}

Aggregate multiple VariableDomains into a contiguous Vector for use by a numerical solver.

Access as a contiguous Vector using copyto!(va::VariableAggregator, x) and copyto!(x, va::VariableAggregator)

PALEOboxes.VariableAggregator — Method

VariableAggregator(vars, cellranges, modeldata)

Create a VariableAggregator for collection of Variables vars, with indices from corresponding cellranges, for modeldata.

cellranges may contain nothing entries to indicate whole Domain.

Defining CellRanges

PALEOboxes.AbstractCellRange — Type

AbstractCellRange

Defines a range of cells within a Domain.

Fields

All implementations should define:

domain::Domain: the Domain covered by this cellrange.

operatorID::Int: If operatorID==0, call all Reactions, otherwise only call those with matching operatorID (this enables operator splitting).

indices: an iterable list of cell indices.

And then may provide subtype-specific fields defining additional ranges of cells.

PALEOboxes.CellRange — Type

CellRange <: AbstractCellRange

Defines a range of cells in a specified Domain as a linear list.

Fields

domain

operatorID

indices

may be any valid Julia indexing range thing eg 1:100, [1 2 3 4], etc

PALEOboxes.CellRangeColumns — Type

CellRangeColumns <: AbstractCellRange

Defines a range of cells in a specified Domain, organised by columns.

Fields

domain

operatorID

indices

iterator through all cells in arbitrary order

columns

iterator through columns: columns[n] returns a Pair icol=>cells where cells are ordered top to bottom

PALEOboxes.create_default_cellrange — Function

create_default_cellrange(model::Model [; operatorID=0]) -> Vector{AbstractCellRange}

Create a Vector of CellRange instances covering the entire model.

PALEOboxes.Grids.create_default_cellrange — Function

create_default_cellrange(domain, grid, [; operatorID=0]) -> CellRange

Create a CellRange for entire domain and supplied operatorID

create_default_cellrange(domain, grid::Nothing [; operatorID=0]) -> CellRange

Create a CellRange for entire domain. Fallback for a domain with no grid

create_default_cellrange(domain, grid::UnstructuredVectorGrid [; operatorID=0]) -> CellRange

Create a CellRange for entire domain - use linear index.

create_default_cellrange(domain, grid::UnstructuredColumnGrid [; operatorID=0]) -> CellRangeColumns

Create a CellRange for entire domain. Return a PB.CellRangeColumns with iterators for columns and cells.

create_default_cellrange(domain, grid::Union{CartesianLinearGrid, CartesianArrayGrid} [; operatorID=0]) -> CellRangeColumns

Create a CellRange for entire domain. Return a PB.CellRangeColumns provided by cellrange_cartesiantile

PALEOboxes.Grids.cellrange_cartesiantile — Function

cellrange_cartesiantile

Create a range of cells within a region of CartesianLinearGrid specified by rangestuple in a specified PB.Domain

rangestuple is a tuple of Cartesian index ranges eg (1:9, :, :) for a 3D grid.

3D case return a CellRangeColumns

return partitioning of a 3D Domain into n tiles

Main loop

PALEOboxes.set_tforce! — Function

set_tforce!(van::VariableAggregatorNamed, val; allow_missing::Bool=true)

Optimised non-allocating version of set_values!(van, :global, :tforce, val)

PALEOboxes.do_deriv — Function

do_deriv(dispatchlists, deltat::Float64=0.0)

Wrapper function to calculate entire derivative (initialize and do methods) in one call. dispatchlists is from create_dispatch_methodlists.

PALEOboxes.dispatch_methodlist — Function

dispatch_methodlist(dl::ReactionMethodDispatchList, deltat::Float64=0.0)

Dispatch to a list of methods.

Implementation

As an optimisation, uses @generated for Type stability and to avoid dynamic dispatch, instead of iterating over lists.

ReactionMethodDispatchList fields are Tuples hence are fully Typed, the @generated function emits unrolled code with a function call for each Tuple element.

Diagnostics

PALEOboxes.show_methods_setup — Function

show_methods_setup(model::Model)

Display ordered list of Reaction setup methods (registered by add_method_setup!, called by dispatch_setup)

PALEOboxes.show_methods_initialize — Function

show_methods_initialize(model::Model)

Display ordered list of Reaction initialize methods (registered by add_method_initialize!, called by do_deriv at start of each model timestep).

PALEOboxes.show_methods_do — Function

show_methods_do(model::Model)

Display ordered list of Reaction do methods (registered by add_method_do!, called by do_deriv for each model timestep).

PALEOboxes.show_variables — Function

show_variables(obj, ...) -> Table

Show all Variables attached to PALEO object obj

PALEOboxes.show_links — Function

show_links(vardom::VariableDomain)
show_links(io::IO, vardom::VariableDomPropDep)

Display all VariableReactions linked to this VariableDomain

PALEOboxes.show_parameters — Function

show_parameters(model) -> DataFrame

Get parameters for all reactions in model.

Examples:

Show all model parameters using VS Code table viewer:

julia> vscodedisplay(PB.show_parameters(run.model))

Write out all model parameters as csv for import to a spreadsheet:

julia> CSV.write("pars.csv", PB.show_parameters(run.model))

show_parameters(react::AbstractReaction) -> DataFrame
show_parameters(classname::AbstractString) -> DataFrame

list all parameters for a Reaction react instance or classname

Accessing model objects

Model

PALEOboxes.get_domain — Function

get_domain(model::Model, name::AbstractString) -> Domain or nothing
get_domain(model::Model, domainid) -> Domain

Get Domain by name (may be nothing if name not matched) or domainid (range 1:num_domains).

PALEOboxes.get_reaction — Method

get_reaction(model::Model, domainname, reactionname) -> Reaction or nothing

Get Reaction by domainname and reaction name

Domains

PALEOboxes.get_variable — Function

get_variable(obj, varname, ...) -> variable

Get variable <: VariableBase by name from PALEO object

PALEOboxes.get_variables — Function

get_variables(varlist::AbstractVarList) -> Vector{VariableReaction}

Return VariableReaction in varlist as a flat Vector.

get_variables(domain; hostdep=nothing, vfunction=VF_Undefined) -> Vector{VariableDomain}

Get domain variables, optionally filtering for subsets based on hostdep and :vfunction attribute

get_variables(domain, filter) -> Vector{VariableDomain}

Get subset of domain variables where filter(var) == true.

get_variables(method::AbstractReactionMethod; filterfn = v -> true) -> Vector{VariableReaction}

Get VariableReactions from method.varlists as a flat Vector, optionally restricting to those that match filterfn

get_variables(reaction, localname) -> Vector{VariableReaction}
get_variables(reaction; [filterfn=v->true]) -> Vector{VariableReaction}

Get matching Variables from all ReactionMethods.

PALEOboxes.get_host_variables — Function

get_host_variables(domain, vfunction; [match_deriv_suffix=""] [, operatorID=0] [, exclude_var_nameroots=[]] [, verbose=false]) 
    -> (host_vars, host_deriv_vars)

Get state Variables with VariableFunction vfunction, and optionally corresponding time derivative with VariableFunction VF_Deriv and name matching hostvarname*<match_deriv_suffix`>.

Optionally filter by operatorID, omit Variables with name matching exclude_var_nameroots.

PALEOboxes.get_unallocated_variables — Function

get_unallocated_variables(domain, modeldata) -> Vector{VariableDomain}

Return any unallocated variables (host-dependent variables which have no data pointer set)

PALEOboxes.get_reaction — Method

get_reaction(domain, reactname) -> Reaction or nothing

Get a reaction by name.

VariableDomain

Low-level access to individual Variables.

PALEOboxes.set_data! — Function

set_data!(var::VariableDomain, modeldata, data)

Set VariableDomain to a Field containing data.

Calls wrap_field to create a new Field.

PALEOboxes.get_data — Function

get_data(var::VariableDomain, modeldata::AbstractModelData) -> field.values
get_data(var::VariableDomain, modeldata::AbstractDomainData) -> field.values

Get Variable var data array from Field.values

get_data(va::VariableAggregator) -> Vector

Allocate Vector and set to values of aggregated Variables va.

PALEOboxes.get_data_output — Function

get_data(var::VariableDomain, modeldata::AbstractModelData) -> get_values_output(field.values)
get_data(var::VariableDomain, modeldata::AbstractDomainData) -> get_values_output(field.values)

Get a sanitized version of Variable var data array for storing as output from [get_values_output]@ref)(Field.values)